rename bert config in allocator
